requirements.txt
google-cloud-aiplatform
google-cloud-storage
google-genai
gitpython
google-adk
python-dotenv
pydantic
a2a-sdk
uvicorn
httpx
deprecated
python-multipart
